rebase
finally finished libpnged, my own personal png decoding library, came
back to this project, (which was the reason for making libpnged in the
first place) and when i went to commit, found out that i had rebased,
and forgot to commit it. fb is jank so i will be taking yet another
break from this project, and go work on libfb. seeya...